# swagger.yaml
openapi: 3.0.3
info:
  title: Navigatorx API
  version: 1.0.0
  description: OpenStreetMap routing engine API for shortest-path routing, alternative routing, map bounds, and client-side real-time map-matching bootstrap data.
  contact:
    name: Lintang Birda Saputra
    email: lintangbirdasaputra23@gmail.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6060
    description: Local development server
tags:
  - name: Routing
    description: Point-to-point routing and alternative route queries.
  - name: Map Matching
    description: Offline map matching for GPX tracks.
  - name: Map Data
    description: Map bounds and tiled graph data for client-side map matching.
  - name: System
    description: Operational endpoints.
paths:
  /api/computeRoutes:
    get:
      tags:
        - Routing
      summary: Compute shortest path
      description: Calculates the shortest path between an origin and destination coordinate.
      parameters:
        - $ref: "#/components/parameters/OriginLat"
        - $ref: "#/components/parameters/OriginLon"
        - $ref: "#/components/parameters/DestinationLat"
        - $ref: "#/components/parameters/DestinationLon"
        - $ref: "#/components/parameters/UseAnnotation"
        - $ref: "#/components/parameters/Reroute"
        - $ref: "#/components/parameters/StartEdgeId"
      responses:
        "200":
          description: Successful routing response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortestPathEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/computeAlternativeRoutes:
    get:
      tags:
        - Routing
      summary: Compute alternative routes
      description: Calculates up to `k` alternative routes between an origin and destination coordinate.
      parameters:
        - $ref: "#/components/parameters/OriginLat"
        - $ref: "#/components/parameters/OriginLon"
        - $ref: "#/components/parameters/DestinationLat"
        - $ref: "#/components/parameters/DestinationLon"
        - in: query
          name: k
          required: true
          description: Number of alternative routes to request.
          schema:
            type: integer
            minimum: 1
        - $ref: "#/components/parameters/UseAnnotation"
        - $ref: "#/components/parameters/Reroute"
        - $ref: "#/components/parameters/StartEdgeId"
      responses:
        "200":
          description: Successful alternative-routes response. The route list can be empty when no alternatives are available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlternativeRoutesEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/mapmatching:
    post:
      tags:
        - Map Matching
      summary: Offline map matching
      description: Performs offline map matching on a GPX track. Accepts a GPX XML body and returns matched GPS points along with the route path.
      parameters:
        - in: query
          name: gps_radiuses
          required: false
          description: Semicolon-separated per-point GPS accuracy radii in meters. The count must match the number of trackpoints in the GPX body.
          schema:
            type: string
          example: "10;15;10"
        - in: query
          name: gps_radius
          required: false
          description: Single GPS accuracy radius in meters applied to all points. Ignored when `gps_radiuses` is provided.
          schema:
            type: number
            format: double
            minimum: 0
          example: 10
        - in: query
          name: include_route_path
          required: false
          description: Whether to include the reconstructed route path in the response.
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        description: GPX XML track to match.
        content:
          application/gpx+xml:
            schema:
              $ref: "#/components/schemas/GPX"
          application/xml:
            schema:
              $ref: "#/components/schemas/GPX"
      responses:
        "200":
          description: Successful offline map-matching response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfflineMapMatchingEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/boundingBox:
    get:
      tags:
        - Map Data
      summary: Get map bounding box
      description: Returns the bounding box of the loaded routing graph.
      responses:
        "200":
          description: Successful bounding-box response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoundingBoxEnvelope"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/tile/{userGeohash}:
    get:
      tags:
        - Map Data
      summary: Get map-matching graph tile
      description: Returns an S2-compressed graph tile for the requested precision-6 geohash cell plus its eight neighboring cells. Clients decompress the response with the S2 framing format and rebuild their local map-matching graph from the newline-delimited edge records.
      parameters:
        - in: path
          name: userGeohash
          required: true
          description: Geohash of the user's current cell.
          schema:
            type: string
            minLength: 1
          example: qq7wqd
      responses:
        "200":
          $ref: "#/components/responses/MapMatchingTile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Tile file not found.
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/tile-init:
    get:
      tags:
        - Map Data
      summary: Initialize client-side map matching
      description: Returns the number of vertices in the loaded road-network graph so clients can initialize their local map-matching graph.
      responses:
        "200":
          description: Successful map-matching initialization response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TileInitEnvelope"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/tile-init-transition-matrix:
    get:
      tags:
        - Map Data
      summary: Get map-matching transition matrix
      description: Returns the transition-count matrix file used by client-side real-time map matching.
      responses:
        "200":
          description: Transition matrix text file.
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Transition matrix file not found.
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /healthz:
    get:
      tags:
        - System
      summary: Health check
      description: Returns a plain-text heartbeat while the server is ready to receive traffic.
      responses:
        "200":
          description: Server is healthy.
          content:
            text/plain:
              schema:
                type: string
                example: .
        "503":
          description: Server is shutting down.
          content:
            text/plain:
              schema:
                type: string
                example: Shutting down
components:
  parameters:
    OriginLat:
      in: query
      name: origin_lat
      description: Latitude of the origin point.
      required: true
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    OriginLon:
      in: query
      name: origin_lon
      description: Longitude of the origin point.
      required: true
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    DestinationLat:
      in: query
      name: destination_lat
      description: Latitude of the destination point.
      required: true
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    DestinationLon:
      in: query
      name: destination_lon
      description: Longitude of the destination point.
      required: true
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    UseAnnotation:
      in: query
      name: useAnnotation
      description: Set to `true` to include per-edge annotation details inside each driving direction.
      required: false
      schema:
        type: boolean
        default: false
    Reroute:
      in: query
      name: reroute
      description: Set to `true` when the request is a reroute. If this parameter is present, `start_edge_id` must also be provided.
      required: false
      schema:
        type: boolean
        default: false
    StartEdgeId:
      in: query
      name: start_edge_id
      description: Unsigned road-segment ID where the user is currently located. Required when `reroute` is present.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
  responses:
    MapMatchingTile:
      description: S2-compressed graph tile. The decompressed payload is newline-delimited text where each line is one map-matching edge record in the format `edge_id tail_vertex_id head_vertex_id length geometry_count lat_0 lon_0 ... lat_n lon_n`.
      content:
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/MapMatchingTile"
    BadRequest:
      description: Bad request or validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: No route could be found from the origin coordinate to the destination coordinate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServiceUnavailable:
      description: Server is shutting down or unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ShortestPathEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/ShortestPathResponse"
    AlternativeRoutesEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AlternativeRoutesResponse"
    BoundingBoxEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/BoundingBoxResponse"
    TileInitEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/TileInitResponse"
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: Bad Request
            message:
              type: string
              example: origin_lat is required and must be a valid float
    ShortestPathResponse:
      type: object
      required:
        - travel_time
        - path
        - distance
        - driving_directions
      properties:
        travel_time:
          type: number
          format: double
          description: Total travel time for the route in minutes.
        path:
          type: string
          description: Encoded Google polyline string representing the path geometry.
        distance:
          type: number
          format: double
          description: Total distance of the route in meters.
        driving_directions:
          type: array
          description: Driving directions and maneuvers for the route.
          items:
            $ref: "#/components/schemas/DrivingDirection"
    AlternativeRoutesResponse:
      type: object
      required:
        - alternative_routes
      properties:
        alternative_routes:
          type: array
          description: Alternative routes ordered by the routing engine.
          items:
            $ref: "#/components/schemas/ShortestPathResponse"
    BoundingBoxResponse:
      type: object
      required:
        - min_lat
        - min_lon
        - max_lat
        - max_lon
      properties:
        min_lat:
          type: number
          format: double
          description: Minimum latitude of the map boundary.
        min_lon:
          type: number
          format: double
          description: Minimum longitude of the map boundary.
        max_lat:
          type: number
          format: double
          description: Maximum latitude of the map boundary.
        max_lon:
          type: number
          format: double
          description: Maximum longitude of the map boundary.
    TileInitResponse:
      type: object
      required:
        - number_of_vertices
      properties:
        number_of_vertices:
          type: integer
          description: Number of vertices in the loaded road-network graph.
          minimum: 0
    MapMatchingTile:
      type: string
      format: binary
      description: S2-compressed tile file generated by `MapAttributesEngine.PreprocessTiles`. The uncompressed stream contains one `MapTileEdgeRecord` per line for every edge in the requested geohash cell and every edge in its neighboring cells.
    MapTileEdgeRecord:
      type: object
      description: Logical structure of one decompressed tile line. The wire format is whitespace-delimited text, not JSON.
      required:
        - edge_id
        - tail_vertex_id
        - head_vertex_id
        - length
        - geometry_count
        - geometry
      properties:
        edge_id:
          type: integer
          format: int32
          minimum: 0
          description: Road-network edge ID.
        tail_vertex_id:
          type: integer
          format: int32
          minimum: 0
          description: Tail vertex ID of the edge.
        head_vertex_id:
          type: integer
          format: int32
          minimum: 0
          description: Head vertex ID of the edge.
        length:
          type: number
          format: double
          minimum: 0
          description: Edge length in meters.
        geometry_count:
          type: integer
          minimum: 0
          description: Number of coordinate pairs that follow in the record.
        geometry:
          type: array
          description: Edge geometry coordinates, serialized on the wire as repeated `lat lon` pairs.
          items:
            $ref: "#/components/schemas/Coordinate"
    DrivingDirection:
      type: object
      required:
        - instruction
        - turn_point
        - street_name
        - travel_time
        - distance
        - turn_bearing
        - turn_type
        - suggest_alternatives
      properties:
        instruction:
          type: string
          description: Human-readable driving instruction.
        annotation:
          $ref: "#/components/schemas/Annotation"
        turn_point:
          description: Coordinate where the maneuver occurs.
          $ref: "#/components/schemas/Coordinate"
        street_name:
          type: string
          description: Name of the street the user is navigating on.
        travel_time:
          type: number
          format: double
          description: Travel time for this instruction in minutes.
        distance:
          type: number
          format: double
          description: Distance covered by this instruction in meters.
        turn_bearing:
          type: number
          format: double
          description: Bearing or heading at the maneuver point.
        turn_type:
          type: string
          description: Turn classification or turn-table ID.
        suggest_alternatives:
          type: boolean
          description: Whether the client should suggest alternative routes at this maneuver.
    Annotation:
      type: object
      description: Per-edge route annotation. Arrays are populated when `useAnnotation=true`.
      properties:
        duration:
          type: array
          description: Per-edge travel durations in seconds.
          items:
            type: number
            format: double
        distance:
          type: array
          description: Per-edge distances in meters.
          items:
            type: number
            format: double
        geometry:
          type: string
          description: Encoded polyline geometry for the annotated route.
        edge_ids:
          type: array
          description: Road-segment IDs for each edge in the annotation.
          items:
            type: integer
            format: int32
            minimum: 0
        edge_geometry_offset:
          type: array
          description: Geometry offsets for each edge in the annotation.
          items:
            type: integer
            format: int32
            minimum: 0
    Coordinate:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: double
          description: Latitude coordinate.
        lon:
          type: number
          format: double
          description: Longitude coordinate.
    GPX:
      type: object
      description: GPX XML document containing a track with trackpoints.
      properties:
        trk:
          type: object
          properties:
            trkseg:
              type: object
              properties:
                trkpt:
                  type: array
                  items:
                    $ref: "#/components/schemas/TrkPt"
    TrkPt:
      type: object
      description: A single trackpoint within a GPX track segment.
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: Latitude of the trackpoint.
        lon:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: Longitude of the trackpoint.
        time:
          type: string
          format: date-time
          description: Timestamp of the trackpoint in RFC 3339 format.
    OfflineMapMatchingEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/OfflineMapMatchingResponse"
    OfflineMapMatchingResponse:
      type: object
      required:
        - matched_points
        - route_path
      properties:
        matched_points:
          type: array
          description: GPS points matched to the road network.
          items:
            $ref: "#/components/schemas/MatchedGPSPoint"
        route_path:
          type: array
          description: Coordinates of the reconstructed route path.
          items:
            $ref: "#/components/schemas/Coordinate"
    MatchedGPSPoint:
      type: object
      required:
        - gps_point
        - edge_id
        - matched_coord
        - predicted_gps_coord
        - edge_initial_bearing
      properties:
        gps_point:
          $ref: "#/components/schemas/GPSPoint"
        edge_id:
          type: integer
          format: int32
          minimum: 0
          description: Road-segment ID matched to this GPS point.
        matched_coord:
          $ref: "#/components/schemas/Coordinate"
        predicted_gps_coord:
          $ref: "#/components/schemas/Coordinate"
        edge_initial_bearing:
          type: number
          format: double
          description: Initial bearing of the matched edge in degrees.
    GPSPoint:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: double
          description: Latitude of the GPS point.
        lon:
          type: number
          format: double
          description: Longitude of the GPS point.
        time:
          type: string
          format: date-time
          description: Timestamp of the GPS reading.
        speed:
          type: number
          format: double
          minimum: 0
          description: Speed at this GPS point in m/s.
        delta_time:
          type: number
          format: double
          minimum: 0
          description: Time delta in seconds from the previous GPS point.